# api.yaml
openapi: 3.0.3
info:
  title: Grocery App - OpenAPI 3.0
  description: |-
    This is the api documentation for the grocery store application. as part of the Modern Application Development 1 Project by Nivedhaa Srikanth . 
  version: 1.0.11
servers:
  - url: http://127.0.0.1:5000/
...
paths:
  /products:
    get:
      tags:
        - product
      summary: Get all the stores products
      description: Get all the stores products as a list of dictionaries
      operationId: getProduct
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        '405':
          description: Invalid input
    post:
      tags:
        - product
      summary: Add a new product to the store
      description: Add a new product to the store
      operationId: addProduct
      requestBody:
        description: Create a new product in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
          application/xml:
            schema:
              $ref: '#/components/schemas/Product'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        '409':
          description: Product with this name already exists
...
  /category:
    post:
      tags:
        - category
      summary: Add a new category
      description: Adds a new category
      operationId: addCategory
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Category name is required and should be a string
        '409':
          description: Category already exists
...
components:
  schemas:
    Product:
      required:
        - name
        - product_id
        - price
        - category_id
        - description
      type: object
      properties:
        id:
          type: integer
          example: 10
        name:
          type: string
          example: amul butter
        price:
          type: number
          example: 280.4
        desctiption:
          type: string
          example: amul butter
        category_id:
          $ref: '#/components/schemas/Category'
        stock:
          type: integer
          example: 10
          
      xml:
        name: Product
    Category:
      type: object
      properties:
        category_id:
          type: integer
          example: 1
        name:
          type: string
          example: Fruits and Vegetables
      xml:
        name: category
...